List svelte and final-form as "external" dependencies in rollup config
... so that it leaves it as `import from 'svelte'` instead of including the contents of svelte into
this library's bundle.

(See: https://rollupjs.org/guide/en/#external)

The consumer of this library will already have svelte and final-form as dependencies and we don't
want to end up with 2 copies of svelte or final-form. Listed them under peerDependencies
to make sure it warns the consumer if they are not installed in their app.