Remove reference to process.env.NODE_ENV, since it won't exist in the browser
Otherwise this will cause an error when this code is run in the browser (unless the consumer uses a
bundler that replaces that reference, which we can't be sure that they will).

It looks like this was left over from when some code was copied from react-final-form.

And it would be fine if we also copied over the corresponding part from their rollup.config.js. But
react-final-form's rollup.config.js is a lot more complex than ours and I'm not sure we want to add
everything it had, without a good reason. They had a 'minify' build-time config option to let them
build both a .min.js (which didn't include this run-time error check) and a regular build (which did
include it). They also left the process.env.NODE_ENV in there when built for UMD (which is assumed
to be Node and *not* a browser). But we currenly we only have a non-minified version and I don't
think we're really targetting Node as a run-time environment, so I'm not convinced that this check
is useful/needed at all.

We could also replace process.env.NODE_ENV with whatever value (when checked in rollup.config.js
during build) the developer who is running `npm run build` for this library may have in their
environment. This has the advantage of being simpler than react-final-form's config, and appears to
be what they do in some other libraries like svelte-material-ui. But:

1. That doesn't really affect the average *consumer* of this library at all (who will simply install
  the version that is published on npm, with no way to influence that NODE_ENV variable since it is
  replaced at build time and not run time)

2. That introduces the risk that a developer who is building and releasing a new package may have
  change their local NODE_ENV environment variable without even realizing it, and without realizing
  that it caused that error check to be _removed_ from the package, for all users.

So until we have a need for it, and a plan for how to use NODE_ENV, I think it's better to just not worry
about it.